+++ src/entity_handler/enemy_type.rs
@@ -1,6 +1,6 @@
 
 
-#[derive(Clone, Copy, PartialEq, Eq, Hash)]
+#[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
 pub enum EnemyType{
     Circle,
     Triangle,